script/requirements.txt
pyspark>=3.0.0
kafka-python>=2.0.2
boto3>=1.17.113
confluent-kafka>=1.7.0
